Ticket #4953 - Changed some IDs to bigint - no upgrade script for this change yet
git-svn-id: file:///data/svn/enum/db/trunk@15981 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ sql/history_base.sql
@@ -4,7 +4,7 @@ CREATE TABLE History (
         valid_from TIMESTAMP NOT NULL DEFAULT NOW(),
         valid_to TIMESTAMP,
         next INTEGER,
-        request_id INTEGER
+        request_id BIGINT 
 );
 
 COMMENT ON TABLE history IS
